#28 Fix: JavaScriptExecutor is throwing StaleElementReferenceException
This occurred because of the changes introduced as part of #27 where
after executing an @Action, action end is logged. In the failure
scenario, the WebElements passed on to the js executor is read from
aspectj proceeding point cut for logging which causes Selenium to look
for the element again in the DOM. This results in stale element
reference exception as the @Action is already performed and the element
could have detached from the DOM.

Solution for this problem works by checking the type of parameters
passed on to the @Action methods. If the parameters are of type
WebElement or WebDriver, framework logs "webelement" and "webdriver"
respectively for the method parameter values instead of using the
toString() method.

In addition to this Retry on @Action failure now attempts to
reinitialize the PageObject before retrying the @Action.